fix tests + move to mock
